Create CategoryDropdown container
