+++ src/containers/category-dropdown/CategoryDropdown.styles.tsx
@@ -1,0 +1,18 @@
+export const styles = {
+  addButton: {
+    justifyContent: 'flex-start',
+    pl: '10px',
+    gap: '5px'
+  },
+  labelCategory: {
+    color: 'primary.600',
+    maxWidth: '464px',
+    width: '100%',
+    display: 'flex',
+    flexDirection: 'column',
+    gap: '4px'
+  },
+  divider: {
+    color: 'primary.300'
+  }
+}
