+++ tests/unit/pages/create-or-edit-lesson/CreateOrEditLesson.spec.jsx
@@ -1,19 +1,9 @@
-import { fireEvent, screen, waitFor } from '@testing-library/react'
-import { renderWithProviders, mockAxiosClient } from '~tests/test-utils'
+import { fireEvent, screen } from '@testing-library/react'
+import { renderWithProviders } from '~tests/test-utils'
 
 import CreateOrEditLesson from '~/pages/create-or-edit-lesson/CreateOrEditLesson'
-import { URLs } from '~/constants/request'
-
-const categoriesNamesMock = [
-  { _id: '650c27618a9fbf234b8bb4cf', name: 'New category in resources!' },
-  { _id: '650c27618a9fbf234b8bb4cd', name: 'Category 1' }
-]
 
 describe('CreateOrEditLesson component test', () => {
-  mockAxiosClient
-    .onGet(URLs.resources.resourcesCategories.getNames)
-    .reply(200, categoriesNamesMock)
-
   beforeEach(() => {
     renderWithProviders(<CreateOrEditLesson />)
   })
@@ -62,52 +52,5 @@ describe('CreateOrEditLesson component test', () => {
     )
 
     expect(errorDescription).toBeInTheDocument()
-  })
-
-  it('should choose the category from options list', async () => {
-    const autocomplete = screen.getByRole('combobox')
-
-    expect(autocomplete).toBeInTheDocument()
-    expect(autocomplete.value).toBe('')
-
-    fireEvent.click(autocomplete)
-    fireEvent.focus(autocomplete)
-
-    fireEvent.change(autocomplete, {
-      target: { value: categoriesNamesMock[1].name }
-    })
-
-    fireEvent.keyDown(autocomplete, { key: 'ArrowDown' })
-    fireEvent.keyDown(autocomplete, { key: 'Enter' })
-
-    await waitFor(() => {
-      expect(autocomplete.value).toBe(categoriesNamesMock[1].name)
-    })
-
-    fireEvent.keyDown(autocomplete, { key: 'ArrowDown' })
-    fireEvent.keyDown(autocomplete, { key: 'Enter' })
-
-    expect(autocomplete.value).toBe(categoriesNamesMock[1].name)
-  })
-
-  it('should click on "add button" in options list', async () => {
-    const autocomplete = screen.getByRole('combobox')
-
-    fireEvent.click(autocomplete)
-    fireEvent.focus(autocomplete)
-
-    fireEvent.keyDown(autocomplete, { key: 'ArrowDown' })
-
-    await waitFor(() => {
-      const addButton = screen.queryByText('myResourcesPage.categories.addBtn')
-
-      fireEvent.click(addButton)
-    })
-
-    await waitFor(() => {
-      const newCategory = screen.getByText('myResourcesPage.categories.name')
-
-      expect(newCategory).toBeInTheDocument()
-    })
   })
 })
